% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{run_linreg}
\alias{run_linreg}
\title{Run a linear regression example using TMB}
\usage{
run_linreg(seed = 123)
}
\arguments{
\item{seed}{See \code{\link[=set.seed]{set.seed()}}}
}
\value{
Nothing
}
\description{
Run a linear regression example using TMB
}
