% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal.R
\name{run_minimal}
\alias{run_minimal}
\title{Minimal TMB example, MLE estimation to find the mean and variance of a normally distributed variable}
\usage{
run_minimal()
}
\value{
Nothing
}
\description{
Minimal TMB example, MLE estimation to find the mean and variance of a normally distributed variable
}
